apps/poller/requirements.txt
requests>=2.31,<3
gtfs-realtime-bindings>=1.0,<2
python-dotenv>=1.0,<2
psycopg2-binary>=2.9,<3
